[M1_TR-212] Reorganized file structure
